requirements.txt
asgiref==3.7.2
Django==4.2.1
django-filter==23.2
djangorestframework==3.14.0
gunicorn==20.1.0
Markdown==3.4.3
# psycopg2-binary==2.9.6
# psycopg2==2.9.6
pytz==2023.3
sqlparse==0.4.4
